Updates
I changed the images paths , added more images and videos , added a new feature in the page1 html called the gallery, finally fixed the about section before the footer in the page 1 , fixed minor mistakes in the shop.html. I then proceeded to add more styling in both pages , as well as interactivity to the html pages.